Move statistical functions into new stats module.
